break out timestamp queue utilities
